+++ sql/mysql-up.sql
@@ -1,4 +1,4 @@
-CREATE TABLE `auth_item` (
+CREATE TABLE `yii_rbac_item` (
   `name` varchar(128) NOT NULL PRIMARY KEY,
   `type` varchar(10) NOT NULL,
   `description` varchar(191),
@@ -6,15 +6,15 @@ CREATE TABLE `auth_item` (
   `createdAt` int(11) NOT NULL,
   `updatedAt` int(11) NOT NULL
 );
-CREATE INDEX `idx-auth_item-type` ON `auth_item` (`type`);
-CREATE TABLE `auth_item_child` (
+CREATE INDEX `idx-yii_rbac_item-type` ON `yii_rbac_item` (`type`);
+CREATE TABLE `yii_rbac_item_child` (
   `parent` varchar(128) NOT NULL,
   `child` varchar(128) NOT NULL,
   PRIMARY KEY (`parent`, `child`),
-  FOREIGN KEY (`parent`) REFERENCES `auth_item` (`name`),
-  FOREIGN KEY (`child`) REFERENCES `auth_item` (`name`)
+  FOREIGN KEY (`parent`) REFERENCES `yii_rbac_item` (`name`),
+  FOREIGN KEY (`child`) REFERENCES `yii_rbac_item` (`name`)
 );
-CREATE TABLE `auth_assignment` (
+CREATE TABLE `yii_rbac_assignment` (
   `itemName` varchar(128) NOT NULL,
   `userId` varchar(128) NOT NULL,
   `createdAt` int(11) NOT NULL,