+++ tests/Base/ItemsStorageTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Rbac\Db\Tests\Base;
 
 use Yiisoft\Db\Query\Query;
+use Yiisoft\Rbac\Db\DbSchemaManager;
 use Yiisoft\Rbac\Db\ItemsStorage;
 use Yiisoft\Rbac\ItemsStorageInterface;
 use Yiisoft\Rbac\Tests\Common\ItemsStorageTestTrait;
@@ -23,7 +24,7 @@ abstract class ItemsStorageTest extends TestCase
         $this->traitTestClear();
 
         $itemsChildrenExist = (new Query($this->getDatabase()))
-            ->from(self::ITEMS_CHILDREN_TABLE)
+            ->from(DbSchemaManager::ITEMS_CHILDREN_TABLE)
             ->exists();
         $this->assertSame(false, $itemsChildrenExist);
     }
@@ -36,7 +37,7 @@ abstract class ItemsStorageTest extends TestCase
         $this->traitTestRemove();
 
         $itemsChildren = (new Query($this->getDatabase()))
-            ->from(self::ITEMS_CHILDREN_TABLE)
+            ->from(DbSchemaManager::ITEMS_CHILDREN_TABLE)
             ->count();
         $this->assertSame($this->initialItemsChildrenCount - $initialItemChildrenCount, $itemsChildren);
     }
@@ -46,7 +47,7 @@ abstract class ItemsStorageTest extends TestCase
         $this->traitTestClearPermissions();
 
         $itemsChildrenCount = (new Query($this->getDatabase()))
-            ->from(self::ITEMS_CHILDREN_TABLE)
+            ->from(DbSchemaManager::ITEMS_CHILDREN_TABLE)
             ->count();
         $this->assertSame($this->initialBothRolesChildrenCount, $itemsChildrenCount);
     }
@@ -56,7 +57,7 @@ abstract class ItemsStorageTest extends TestCase
         $this->traitTestClearRoles();
 
         $itemsChildrenCount = (new Query($this->getDatabase()))
-            ->from(self::ITEMS_CHILDREN_TABLE)
+            ->from(DbSchemaManager::ITEMS_CHILDREN_TABLE)
             ->count();
         $this->assertSame($this->initialBothPermissionsChildrenCount, $itemsChildrenCount);
     }
@@ -68,17 +69,17 @@ abstract class ItemsStorageTest extends TestCase
         $this
             ->getDatabase()
             ->createCommand()
-            ->batchInsert(self::ITEMS_TABLE, ['name', 'type', 'createdAt', 'updatedAt'], $fixtures['items'])
+            ->batchInsert(DbSchemaManager::ITEMS_TABLE, ['name', 'type', 'createdAt', 'updatedAt'], $fixtures['items'])
             ->execute();
         $this
             ->getDatabase()
             ->createCommand()
-            ->batchInsert(self::ITEMS_CHILDREN_TABLE, ['parent', 'child'], $fixtures['itemsChildren'])
+            ->batchInsert(DbSchemaManager::ITEMS_CHILDREN_TABLE, ['parent', 'child'], $fixtures['itemsChildren'])
             ->execute();
     }
 
     private function getStorage(): ItemsStorageInterface
     {
-        return new ItemsStorage(self::ITEMS_TABLE, $this->getDatabase());
+        return new ItemsStorage($this->getDatabase());
     }
 }